admin delete functionality

+++ src/DeepMikoto/ApiBundle/Entity/CodingCategory.php
@@ -168,7 +168,7 @@ class CodingCategory
     public function removeUpload()
     {
         $file = $this->getAbsolutePath();
-        if ( $file ) {
+        if ( $file && file_exists( $file ) && is_file( $file ) ) {
             unlink( $file );
         }
     }
