+++ src/DeepMikoto/ApiBundle/Entity/GamingPost.php
@@ -180,7 +180,7 @@ class GamingPost
     public function removeUpload()
     {
         $file = $this->getAbsolutePath();
-        if ( $file ) {
+        if ( $file && file_exists( $file ) && is_file( $file ) ) {
             unlink( $file );
         }
     }
