+++ src/DeepMikoto/ApiBundle/Entity/PhotographyPostPhoto.php
@@ -162,7 +162,7 @@ class PhotographyPostPhoto
     public function removeUpload()
     {
         $file = $this->getAbsolutePath();
-        if ( $file ) {
+        if ( $file && file_exists( $file ) && is_file( $file ) ) {
             unlink( $file );
         }
     }
